+++ resources/sql/queries.sql
@@ -485,9 +485,10 @@ FROM application_event
 ORDER BY id DESC
 LIMIT 1;
 
--- :name add-application-event! :insert
+-- :name add-application-event! :returning-execute :1
 INSERT INTO application_event (appId, eventData)
-VALUES (:application, :eventdata::jsonb);
+VALUES (:application, :eventdata::jsonb)
+RETURNING id, eventData::TEXT;
 
 -- :name upsert-api-key! :insert
 INSERT INTO api_key (apiKey, comment, users, paths)