+++ src/clj/rems/api/services/attachment.clj
@@ -7,8 +7,7 @@
             [rems.db.applications :as applications]
             [rems.db.attachments :as attachments]
             [rems.util :refer [getx]]
-            [ring.util.http-response :refer [ok content-type header]]
-            [ring.util.io :as ring-io])
+            [ring.util.http-response :refer [ok content-type header]])
   (:import [java.io ByteArrayInputStream ByteArrayOutputStream]
            [java.util.zip ZipOutputStream ZipEntry ZipException]))
 
@@ -46,21 +45,20 @@
     (attachments/save-attachment! file user-id application-id)))
 
 (defn zip-attachments [application]
-  (let [zip-input (ring-io/piped-input-stream
-                   (fn [out]
-                     (with-open [zip (ZipOutputStream. out)]
-                       (doseq [metadata (getx application :application/attachments)]
-                         (let [id (getx metadata :attachment/id)
-                               attachment (attachments/get-attachment id)]
-                           ;; we deduplicate filenames when uploading, but here's a
-                           ;; failsafe in case we have duplicate filenames in old
-                           ;; applications
-                           (try
-                             (.putNextEntry zip (ZipEntry. (getx attachment :attachment/filename)))
-                             (.write zip (getx attachment :attachment/data))
-                             (.closeEntry zip)
-                             (catch ZipException e
-                               (log/warn "Ignoring attachment" (pr-str metadata) "when generating zip. Cause:" e))))))))]
-    (-> (ok zip-input)
+  (let [out (ByteArrayOutputStream.)]
+    (with-open [zip (ZipOutputStream. out)]
+      (doseq [metadata (getx application :application/attachments)]
+        (let [id (getx metadata :attachment/id)
+              attachment (attachments/get-attachment id)]
+          ;; we deduplicate filenames when uploading, but here's a
+          ;; failsafe in case we have duplicate filenames in old
+          ;; applications
+          (try
+            (.putNextEntry zip (ZipEntry. (getx attachment :attachment/filename)))
+            (.write zip (getx attachment :attachment/data))
+            (.closeEntry zip)
+            (catch ZipException e
+              (log/warn "Ignoring attachment" (pr-str metadata) "when generating zip. Cause:" e))))))
+    (-> (ok (ByteArrayInputStream. (.toByteArray out))) ;; extra copy of the data here, could be more efficient
         (header "Content-Disposition" (str "attachment;filename=attachments-" (getx application :application/id) ".zip"))
         (content-type "application/zip"))))