+++ src/clj/rems/db/events.clj
@@ -41,7 +41,8 @@
 (defn get-latest-event []
   (fix-event-from-db (db/get-latest-application-event {})))
 
-(defn add-event! [event]
-  (db/add-application-event! {:application (:application/id event)
-                              :eventdata (event->json event)})
-  nil)
+(defn add-event!
+  "Add event to database. Returns the event as it went into the db."
+  [event]
+  (fix-event-from-db (db/add-application-event! {:application (:application/id event)
+                                                       :eventdata (event->json event)})))